requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pycryptodome==3.19.0
requests==2.31.0
schedule==1.2.0
pytz==2023.3